// app copy/container/Tariff/interface.ts
export interface PromotionalSummaryCardOffer {
    /**
     * Offer Price
     */
    offerPrice: number | null;
}

/**
 * Initial state
 */
export interface IInitialState {
    /**
     * Loading flags per API
     */
    loading: {
        getSubscription: boolean;
    };
    /**
     * Error flags per API
     */
    errors: {
        getSubscription: boolean;
    };
    /**
     * Subscription Id
     */
    subscriptionId: RedTariff | YoungTariff | null;
    /**
     * offers
     */
    subscriptionPayload: TariffWithHardwareResponse | null;
    /**
     * promotional Summary Card Offer
     */
    promotionalSummaryCardOffer: PromotionalSummaryCardOffer,
}

/**
 * Promotional Summary Card offer
 */
export interface PromotionalSummaryCardOfferInterface {
     /**
      * Offer Price
      */
     offerPrice: string | null
}

/**
 * Interface of Tariff Properties
 */
export interface StateProps extends Pick<IInitialStateApp, 'isTauschbonus' | 'isTradeIn'>{
    /**
     * Atomic ID
     */
    atomicId: string | null;
    /**
     * Payload
     */
    subscriptionPayload?: TariffWithHardwareResponse | null;
   /**
    * Offers
    */
    offers: TariffWithHardwareOffer[] | null;
    /**
     * Subscription Id
     */
    subscriptionId: RedTariff | YoungTariff | null;
   /**
    * salesChannel
    */
    salesChannel: SalesChannel | null;
    /**
     * Device Id
     */
    deviceId: string | null;
    /**
     * offer
     */
    offer: TariffWithHardwareOffer | null;
    /**
     * Tariff price starting from month 1 (stairway, or EOL)
     * With discounts (if any), otherwise normal tariff price.
     */
    tariffPrice: PriceWithRuntime | null;
    /**
     * Final tariff price after stairway, ending at month 24
     * With discounts (if any), otherwise normal tariff price.
     */
    endTariffPrice: PriceWithRuntime | null;
    /**
     * Data volume starting from month 1 (stairway, or EOL)
     * Includes additional volume via discounts (if any), otherwise normal data volume.
     */
    dataVolume: DataVolume | null;
    /**
     * Final data volume after stairway, ending at month 24
     * Includes additional volume via discounts (if any), otherwise normal data volume.
     */
    endDataVolume: DataVolume | null;
    /**
     * True if there is a stairway price
     */
    hasStairway: boolean;
    /**
     * Strike price (if any)
     */
    strikePrice: PriceWithRuntime | null;
    /**
     * The price the customer has to pay initially (including insurance)
     */
    priceToPay: PriceWithRuntime | null;
    /**
     * The price the customer has to pay after end of stairway (including insurance)
     */
    endPriceToPay: PriceWithRuntime | null;
    /**
     * Device name
     */
    deviceName: string | null;
    /**
     * isRedplusEligible
     */
    isRedplusEligible: boolean;
    /**
     * promotionalSummaryCardOffer
     */
    promotionalSummaryCardOffer: PromotionalSummaryCardOffer;
}